auto/requirements.txt
requests>=2.31.0
google-auth>=2.22.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.0.0
